refactor: Google 데이터 관련 메서드 위치 변경
